+++ src/main/java/xyz/ncookie/stargazer/domain/member/dto/TokenDto.java
@@ -1,4 +1,4 @@
-package xyz.ncookie.stargazer.domain.member.dto.response;
+package xyz.ncookie.stargazer.domain.member.dto;
 
 public record TokenDto(
 	String accessToken,
